+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/annotation/parameter/SupportParameterType.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe.annotation;
+package com.interface21.webmvc.servlet.mvc.tobe.annotation.parameter;
 
 import java.util.Arrays;
 import java.util.List;
